requirements.txt
openai
python-dotenv
pymupdf